[MOV] test_mail: move access tests in their own file
Purpose is to separate tests about content from tests about access. It
eases bugfixing and improving features while keeping other tests stable
through versions.

Task-4320561

X-original-commit: 79e269335d5905acf6e5df360c1fcfbc176f34cc
